Started Working on Enemy Attacking and Defending
Enemies will have a health percentage that they begin defending at (can be set to 0 for no defense). They will then random between 0 and 1, if that number is greater than a certain defense percentage, it will try to block. So the closer the percentage is to 1, the less chance of blocking.